Fix IP anonymization when client ip is an ipv6 address
Before this fix, if a client had an ipv6 address the creation of the
magic link would fail with this error:

```
invalid input syntax for type inet: "0"
```

This is because the original ip anonymization code below assumed the ip was an
ipv4 address. When an ipv6 address was given, the ip address would get
mangled into '0' and subsequently fail the validation of the GenericIPAddressField.

```py
client_ip[:client_ip.rfind('.')+1] + '0'
```

This commit adds support for anonymizing ipv6 addresses. For these addresses
the last 80 bits are zeroed out. This seems to be the recommended practise,
as also used by Google Analytics.